Update deps
Main thing here is iroh to 0.95, since the public relay servers no
longer support 0.90. Involved some minor API tweaks, with the upshot
that "node" no longer makes sense as part of a QFTF FileTransfer struct,
so that's called "addr" now, which is a breaking change (requires
qftf-web to match). But the old version wasn't working anyway (due to
the relay servers) so no harm there, everyone needs to update anyway.